refactor: change dialog folder structure
