+++ client/src/Components/Dialogs/AnnotationDialogs/RenameDialog.tsx
@@ -1,10 +1,10 @@
 import classNames from "classnames";
 import React, {useState} from "react";
 import {Button, Form, Modal} from "react-bootstrap";
-import {Nullable, Setter} from "../../util/types";
-import {isValidPythonIdentifier} from "../../util/validation";
-import "../SelectionView/SelectionView.css";
-import DialogCSS from "./dialog.module.css";
+import {Nullable, Setter} from "../../../util/types";
+import {isValidPythonIdentifier} from "../../../util/validation";
+import "../../SelectionView/SelectionView.css";
+import DialogCSS from "../dialogs.module.css";
 
 interface RenameDialogProps {
     isVisible: boolean
