+++ client/src/Components/Dialogs/MenuDialogs/ImportPythonPackageDialog.tsx
@@ -1,35 +1,43 @@
 import React, {useState} from "react";
 import {Button, Form, Modal} from "react-bootstrap";
-import {Setter} from "../../util/types";
-import "../SelectionView/SelectionView.css";
+import {useHistory} from "react-router-dom";
+import {Setter} from "../../../util/types";
+import "../../SelectionView/SelectionView.css";
 import Dropzone from 'react-dropzone';
-import {isValidJsonFile} from "../../util/validation";
-import DialogCSS from "./dialog.module.css";
-import AnnotationStore from "../../model/annotation/AnnotationStore";
+import {isValidJsonFile} from "../../../util/validation";
+import DialogCSS from "../dialogs.module.css";
+import PythonPackage from "../../../model/python/PythonPackage";
+import {parsePythonPackageJson} from "../../../model/python/PythonPackageBuilder";
+import AnnotationStore from "../../../model/annotation/AnnotationStore";
 
-interface ImportAnnotationFileDialogProps {
+interface ImportPythonPackageDialogProps {
     isVisible: boolean
-    setIsVisible: Setter<boolean>
+    setIsVisible: Setter<boolean>,
+    setPythonPackage: Setter<PythonPackage>
     setAnnotationStore: Setter<AnnotationStore>
+    setFilter: Setter<string>
 }
 
-export default function ImportAnnotationFileDialog(props: ImportAnnotationFileDialogProps): JSX.Element {
+export default function ImportPythonPackageDialog(props: ImportPythonPackageDialogProps): JSX.Element {
 
     const [fileName, setFileName] = useState("");
-    const [newAnnotationStore, setNewAnnotationStore] = useState(new AnnotationStore());
+    const [newPythonPackage, setNewPythonPackage] = useState<PythonPackage>();
+    const history = useHistory();
 
     const close = () => {
         props.setIsVisible(false);
     };
 
     const submit = () => {
-        if (fileName) {
-            props.setAnnotationStore(newAnnotationStore);
-        }
         props.setIsVisible(false);
+        if (newPythonPackage) {
+            props.setPythonPackage(newPythonPackage);
+            props.setFilter("");
+            history.push("/");
+        }
     };
 
-    const onDrop = (acceptedFiles: File[]) => {
+    const slurpAndParse = (acceptedFiles: File[]) => {
         if (isValidJsonFile(acceptedFiles[acceptedFiles.length - 1].name)) {
             if (acceptedFiles.length > 1) {
                 acceptedFiles = [acceptedFiles[acceptedFiles.length - 1]];
@@ -38,10 +46,8 @@ export default function ImportAnnotationFileDialog(props: ImportAnnotationFileDi
             const reader = new FileReader();
             reader.onload = () => {
                 if (typeof reader.result === 'string') {
-                    const readAnnotationJson = JSON.parse(reader.result);
-                    readAnnotationJson["renamings"] = new Map(Object.entries(readAnnotationJson["renamings"]));
-                    readAnnotationJson["enums"] = new Map(Object.entries(readAnnotationJson["enums"]));
-                    setNewAnnotationStore(AnnotationStore.fromJson(readAnnotationJson));
+                    setNewPythonPackage(parsePythonPackageJson(JSON.parse(reader.result)));
+                    props.setAnnotationStore(new AnnotationStore());
                 }
             };
             reader.readAsText(acceptedFiles[0]);
@@ -54,32 +60,34 @@ export default function ImportAnnotationFileDialog(props: ImportAnnotationFileDi
                size={"lg"}
                className={DialogCSS.modalDialog}>
             <Modal.Header closeButton>
-                <Modal.Title>Import annotation file</Modal.Title>
+                <Modal.Title>
+                    Import Python package
+                </Modal.Title>
             </Modal.Header>
             <Modal.Body>
                 <Form noValidate>
                     <Modal.Body>
                         <Form.Group>
                             <Form.Label>
-                                Select an annotation file to upload.
+                                Select a Python package to upload.
                             </Form.Label>
                             <div className={DialogCSS.dropzone}>
-                                <Dropzone onDrop={onDrop}>
+                                <Dropzone onDrop={slurpAndParse}>
                                     {({getRootProps, getInputProps}) => (
                                         <section>
                                             <div {...getRootProps()}>
-                                                <input {...getInputProps()}/>
+                                                <input {...getInputProps()} />
                                                 <p className={DialogCSS.dropzoneText}>
-                                                    Drag and drop an annotation file here or click to select the
+                                                    Drag and drop a Python package here, or click to select the
                                                     file.<br/>
-                                                    (only *.json will be accepted)
+                                                    (Only *.json will be accepted.)
                                                 </p>
                                             </div>
                                         </section>
                                     )}
                                 </Dropzone>
                             </div>
-                            {fileName && <div><strong>Imported file: </strong>{fileName}</div>}
+                            {fileName && <div><strong>Imported package name: </strong>{fileName}</div>}
                         </Form.Group>
                     </Modal.Body>
                     <Modal.Footer>
