+++ client/src/Components/SelectionView/ParameterNode.tsx
@@ -5,8 +5,8 @@ import AnnotationStore from "../../model/annotation/AnnotationStore";
 import PythonEnum from "../../model/python/PythonEnum";
 import PythonParameter from "../../model/python/PythonParameter";
 import {Nullable, Setter} from "../../util/types";
-import RenameDialog from "../Dialog/RenameDialog";
-import EnumDialog from "./Dialogues/EnumDialog";
+import RenameDialog from "../Dialogs/AnnotationDialogs/RenameDialog";
+import EnumDialog from "../Dialogs/AnnotationDialogs/EnumDialog/EnumDialog";
 import DocumentationText from "./DocumentationText";
 import AnnotationView from "./AnnotationView";
 import "./SelectionView.css";
